migrate events detail page
